requirements.txt
# Core dependencies
numpy>=1.26.0
pandas>=2.1.0
python-dotenv>=1.0.0

# Text Knowledge - Embeddings and Vector Store
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
torch>=2.1.0
transformers>=4.35.0

# Knowledge Graph
networkx>=3.1
rdflib>=7.0.0

# API and Web Access
requests>=2.31.0
biopython>=1.81
beautifulsoup4>=4.12.0
lxml>=4.9.0

# LLM Integration
openai>=1.3.0
anthropic>=0.7.0
huggingface-hub>=0.19.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0.0
scikit-learn>=1.3.0

# PDF Knowledge
pypdf>=3.17.0
